chore: fix build by forcing single file output
